requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cassandra-driver==3.28.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
exceptiongroup==1.1.2
fastapi==0.100.1
fastapi-pagination==0.12.8
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
numpy==1.24.4
pandas==2.0.3
paramiko==3.3.1
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
